Add rake:debug task to quickly build the CSS
Previously to ensure that changes have been working it was required to
build the gem and then push it to a sample application (e.g. Rails &
Compass).

The rake:debug task uses plain Sass - as we do for testing, and creates
the file 'debug.css' without having to rebuild the Gem and fire up an
application.