CF-372: Add sidebar for Notes
This commit adds the sidebar for notes with JS to detect the width
of the sidebar on resize so the fixed action header above the note
'sticks'. The collapse button is also moved back to the bottom of
the sidebar.

CF-372: Remove debug line

CF-372: Add 'currently viewing' section to sidebar

This commit adds the 'currently viewing' section with the user or
location picture, name and title if applicable.